[Refactor] 디렉토리 정리 및 파일 이름 수정
