+++ MateRunner/MateRunner/Data/Network/DataMapping/EmojiFirestoreDTO.swift
@@ -1,5 +1,5 @@
 //
-//  EmojisDTO.swift
+//  EmojiFirestoreDTO.swift
 //  MateRunner
 //
 //  Created by 전여훈 on 2021/11/24.
