+++ MateRunner/MateRunner/Data/Network/DataMapping/RunningResultFirestoreDTO.swift
