+++ MateRunner/MateRunner/Domain/Model/PresonalTotalRecord.swift
@@ -1,0 +1,14 @@
+//
+//  PresonalTotalRecord.swift
+//  MateRunner
+//
+//  Created by 전여훈 on 2021/11/24.
+//
+
+import Foundation
+
+struct PresonalTotalRecord {
+    let distance: Double
+    let time: Int
+    let calorie: Double
+}
