+++ MateRunner/MateRunner/Domain/Model/UserData.swift
@@ -1,0 +1,19 @@
+//
+//  UserData.swift
+//  MateRunner
+//
+//  Created by 전여훈 on 2021/11/24.
+//
+
+import Foundation
+
+struct UserData {
+    let nickname: String
+    let image: String
+    let time: Int
+    let distance: Double
+    let calorie: Double
+    let height: Double
+    let weight: Double
+    let mate: [String]
+}
