+++ MateRunner/MateRunner/Domain/Model/UserProfile.swift
@@ -1,0 +1,14 @@
+//
+//  UserProfile.swift
+//  MateRunner
+//
+//  Created by 전여훈 on 2021/11/24.
+//
+
+import Foundation
+
+struct UserProfile {
+    let iamge: String
+    let height: Double
+    let weight: Double
+}
